+++ src/hx/cppia/CppiaCompiler.cpp
@@ -189,6 +189,12 @@ int sCtxReg = SLJIT_S0;
 int sFrameReg = SLJIT_S1;
 int sThisReg = SLJIT_S2;
 
+// Called from jitted prologues when the script stack limit is hit
+static void cppiaJitStackOverflow(CppiaCtx *inCtx)
+{
+   hx::Throw( HX_CSTRING("Stack Overflow") );
+}
+
 JitReg sJitCtx(SLJIT_S0,jtPointer);
 JitReg sJitFrame(SLJIT_S1,jtPointer);
 JitReg sJitThis(SLJIT_S2,jtPointer);
@@ -196,6 +202,7 @@ JitReg sJitThis(SLJIT_S2,jtPointer);
 
 JitVal sJitCtxFrame = sJitCtx.star(jtPointer, offsetof(CppiaCtx,frame));
 JitVal sJitCtxPointer = sJitCtx.star(jtPointer, offsetof(CppiaCtx,pointer));
+JitVal sJitCtxStackEnd = sJitCtx.star(jtPointer, offsetof(CppiaCtx,stackEnd));
 
 static double sZero = 0.0;
 
@@ -324,7 +331,13 @@ public:
       {
          move( sJitFrame, sJitCtxFrame );
          if (makesNativeCalls)
+         {
              add( sJitCtxPointer, sJitFrame, maxFrameSize );
+             // Deep jit-to-jit recursion bypasses the interpreter's check
+             JumpId stackOk = compare( cmpP_LESS, sJitCtxPointer, sJitCtxStackEnd, 0 );
+             callNative( (void *)cppiaJitStackOverflow, sJitCtx.as(jtPointer) );
+             comeFrom(stackOk);
+         }
       }
 
       if (usesThis)
@@ -808,6 +821,9 @@ public:
    void setMaxPointer()
    {
       add( sJitCtxPointer, sJitFrame, maxFrameSize );
+      JumpId stackOk = compare( cmpP_LESS, sJitCtxPointer, sJitCtxStackEnd, 0 );
+      callNative( (void *)cppiaJitStackOverflow, sJitCtx.as(jtPointer) );
+      comeFrom(stackOk);
    }
 
    void makeAddress(const JitVal &outAddress, const JitVal &inSrc)
@@ -938,7 +954,9 @@ public:
          switch(inSrcType)
          {
             case etInt:
-               emit_fop1( SLJIT_CONV_F64_FROM_S32, inTarget.as(jtInt), inSrc.as(jtFloat) );
+               // The destination of int->float is the FLOAT operand - the swapped
+               // tags made register accounting attribute it to the wrong class
+               emit_fop1( SLJIT_CONV_F64_FROM_S32, inTarget.as(jtFloat), inSrc.as(jtInt) );
                break;
 
             case etObject:
@@ -978,7 +996,7 @@ public:
          switch(inSrcType)
          {
             case etFloat:
-               emit_fop1( SLJIT_CONV_S32_FROM_F64, inTarget.as(jtFloat), inSrc.as(jtInt) );
+               emit_fop1( SLJIT_CONV_S32_FROM_F64, inTarget.as(jtInt), inSrc.as(jtFloat) );
                break;
 
             case etObject: